requirements.txt
pandas
numpy
matplotlib
seaborn
statsmodels
scipy
textblob
pyyaml
requests
